Make sure that Hash[Int].new gets the correct descriptor
According to a bisect, this issue has existed since at least Christmas.
The problem was that Hash.new would do nothing about a descriptor.  Before
the scalar refactor, that meant that the descriptor was nqp::null.  After
that refactor, it has a default descriptor, but still not the right thing.
Apparently, when you say "my Int %h", the correct descriptor is codegenned
or otherwise correctly created.

This commit introduces a multi for Map.new: one without parameters, and
one with.  The one with calls the .new without parameters.  For the TypedHash
roles, there is a no-argument .new candidate that creates the correct
descriptor.  Which fixes R#2348.