Streamline MAIN_HELPER and add some features
If no appropriate MAIN candidate was found, it would generate usage
info for *all* candidates.  With this change, it will only generate
usage info for those candidates that match with the given *positional*
parameters only.  This will cause "zef install" to only show the help
for the "install".

Also, if <named-anywhere> is active, then the named parameters will
be shown *after* the positional parameters in any usage information.

Inspired by the October Amsterdam.PM meeting