Refactor MAIN_HELPER / add support for &ARGS-TO-CAPTURE
This introduces support for an &ARGS-TO-CAPTURE to exist in the code
where the MAIN also exists.  It takes an array of ARGS, and is supposed
to return a Capture.  If there is no such sub, then it will use the
default one living inside MAIN_HELPER.

This should allow GetOpt::Long type of modules to simplify their code
to just supply an &ARGS-TO-CAPTURE sub without having to worry about
the actual dispatch specifics.