Don't unconditionally add parameterizations to SC
A parameterization of a type may not create a new type, but instead
return an already produced type. Thus, it may already belong to the
SC of another compilation unit. Placing it into ours results in a
duplicate of the object - in this case a type object - existing after
deserialization.

This was the cause of issue #2400: the type check failed because the
duplicate of the type object was quite rightly not in the MRO of the
meta-object.

The fix is simply to only add it if it's not already in an SC. If it
is, we need do nothing more.