Do a lot less work upon a redeclaration
The variable is already there, so we don't need to go and do all of
the things we would were it not. This also avoids scribbling over the
things the optimizer uses to tell the difference between a parameter
or declared variable, fixing an issue introduced by the lexical to
local lowering.