Make Date.succ/pred and Date ± N about 3x as fast (usually)
This should make iterating over a range of Dates or a loop with ++$date
about 3x as fast.  Optimization based on the observation that if only the
date value can change, we just need to clone the object and update the day
value.  Complicating factor could be the daycount attribute that also needs
updating if it is set.  If anything apart the day needs to change, then it
will fallback to the old logic, creating a new object using the daycount
attribute.