Add support for Pair >>op<< Pair
Since Pairs are generally considered to be 1 element hashes, use the same
logic.  In the situation where if they would be hashes and no matching
keys, an empty hash would be returned.  Since there is no such thing as an
empty Pair, Nil is returned, rather than Nil => Nil.

This commit breaks some spectests because of R#2484.  The next commit works
around that.  Splitting into two separate commits so that the next commit
can simply be reverted once R#2484 is fixed.