We need to cache on name
nqp::objectid is different for each call (not sure why), so that meant
that the cache was filled with stale data.  Now, if an op requests an
META_ASSIGN, and it has no name, then it will take a slower path very
close to what it originally was.  If it *does* have a name, then it will
be cached on the name of the original op.