Fix is-lazy propagation bug in List::Reifier
Triangle reduce was not recognizing Lists containing lazy Slips as lazy.
This can be tracked back to the is-lazy method on List::Reifier (and was
ultimately being propagated up to the triangle reduce routine). is-lazy
in List::Reifier was not considering the possibility of lazy Slips. Fix
this by having the method check any $!future elements it may have for
laziness.

See [Github Issue #2122](https://github.com/rakudo/rakudo/issues/2122)