Give 6.d correct `is rw` Proxy semantics
We had a long-standing bug where Proxy would not be decontainerized
when returning from a routine. This behavior should only apply for `is
rw` routines, but accidentally applied to all of them. This was fixed
while changing decontrv handling to use a spesh plugin, but it turned
out some modules depended on the wrong behavior, so the bug was then
emulated in the spesh plugin.

With the first 6.d-supporting release, we'll take the opportunity to
have the fix in 6.d, while compiling programs doing `use v6.c` to use
the previous buggy behavior.

This resolves issue #2113.