Avoid repeated PseudoStash creation in regexes
If we do something like:

        my $r = q|\d+|;
        "asdf123fdsa321" ~~ /<$r>/ for ^300_000;

Then we would previously created 5 PseudoContext objects to pass to
INTERPOLATE_MATCH, one for each position we visited in the string. That
is clearly wasteful: one per time the regex is invoked will do. So, do
that. This saves about 20% of runtime in the code sample above, meaning
that #2537 has now won back more than the 10%-15% it was reported to
have regressed by previously.