Fix missing redeclaration of bound outer error
The code `my $a = 42; { say $a; my ($a) }` would fail to give an error
that `$a` cannot be redeclared because it is already bound. The error
was correctly reported in the non-signature case of a declaration (e.g.
`my $a = 42; { say $a; my $a }`). This makes sure it reports in both
cases, fixing #2570.