Restore CALLER:: behavior on non-dynamic vars
We retain the lexical slot, but fill it with something that will always
come out as non-dynamic in the PseudoStash check, thus restoring the
behavior expected in the spectests. It'd be nice to be able to fully do
away with this at some point, but that needs either a better mechanism
or a spec change.