Don't destroy the standard QAST::Block structure
We have an optimization that discards a QAST::Stmts that wraps a single
statement. However, we also rely on the standard structure of every
QAST::Block having two QAST::Stmts nodes within in it, and optimizations
are refused in some places if that isn't found. So, never do this at the
top block level.