Don't mark $_ container with `is dynamic` in 6.d
We already changed the spectests to allow for this change to happen
during the 6.d release preparations.

The moving of `finishpad` is to make sure that we don't create the topic
container descriptor until after the language version is decided;
otherwise, even in 6.c it will be created as non-dynamic and then
re-used after that point in that mode.