Differentiate required and optional implicit topic
When we have `for @a { }` then there is no need for the `$_` parameter
to be optional, since we know we'll always be invoking the block with a
parameter. This should lead to slightly more compact/efficient code for
such cases, and will make the static optimizer's job easier with regard
to having to try and track outer lexical accesses.