Introducing $*FULL-PRECISION
By default False.  When set to True, it will *not* downgrade a Rat to a
Num (float) but instead upgrade it to a FatRat.

This is an alternate approach to "use full-precision": whereas the pragma
only functions in the lexical scope it is used, using the dynamic variable
has the advantage of being applicable to external libraries that use Rats
as well.

Since the dynamic variable is only checked if a decision needs to be made
about downgrading or upgrading, this should not affect normal Rat (or FatRat)
operation in any way.