Start to lower $_ from lexical to local
This almost passes spectest, thanks to the various preparatory changes
that were done. It also gives a quite nice speedup on the empty range
benchmark, since the unused `$_` now uses a local, which we can in turn
see is completely unused, and therefore throw out a boxing operation,
thus meaning it can JIT into a rather tighter loop. Alas, it seems that
despite the wins, this costs us something on some benchmarks, which will
need further examination.