Empty the unused QAST::Block if { ... } is a hash
We take the contents of the QAST::Block and use it as the hash.
However, we then failed to clear out the entries inside of that
QAST::Block. This means we would compile that part of the tree twice,
which is both a waste of compilation time and a bunch of bytecode we'll
never run. Furthermore, it turns out this tree duplication in separate
blocks could confuse the optimizer (which is how I discovered it).