Poison lowering if we see a quasi quote
It might refer to variables outside of itself, and we can't yet easily
track if that is the case. Macros are used very little, and will
probably be completely re-done based on the 007 work, so for now just do
the easiest thing possible to fix things like OO::Monitors that make use
of them.