More aggressive unused $_ deletion
Now that $_ is lexical (at least, under 6.d), we can in most cases
better understand its usage, and delete it when we see there is no use
of it at all. This recovers the performance we lost in the lowering of
$_, caused by the fact that the reliance on lazy lexical initialization
went away, and we were thus always initializing it.