Make `does X::Control` imply throwing to CONTROL
Previously, there was no way to define custom control exceptions. This
provides a way to do so: an exception that `does X::Control` will be
thrown as a control exception, and so seen by CONTROL blocks rather than
CATCH ones.

This space needs a bit more fleshing out generally; none of the existing
CX classes can actually be thrown correctly and just go as CATCH
exceptions (this was already the situation; this change does not make
things worse).