Fix hang in Channel subscribe/close/drain
When we polled a Channel, we forgot to re-insert the CLOSE or FAIL
message. This meant that under certain use patterns, such as when we
had a subscription to a Channel, closed it, and then tried to drain it
with a `for` loop, we could get a hang, since the draining loop never
saw the CLOSE message.