Fix Inline::Perl5 failing to load when installed through the Staging repository
Marking the %quote_lang_cache as neverrepossess was not enough. It gets
serialized anyway and the MVM_CF_NEVER_REPOSSESS on the object doesn't survive
serialization/deserialization. This caused references to
CompUnit::Repository::Staging's serialization context to end up in the installed
Inline::Perl5::Parent when they shouldn't have. And That's why we ended up with
missing dependencies when loading Inline::Perl5 from an RPM package.

Fix by turning the lexical variable into an attribute of $*W so it won't get
serialized at all. This should also help multi threaded compilation (EVAL).

Thanks to jnthn++ for suggesting the implemented solution!