Add a `dyanmic-scope` lexical pragma
This allows for specifying that:

* Without args, all variables declared in the scope the pragma applies
  to will be automatically `is dynamic`
* With args, the named variables declared in the scope the pragma
  applies to will be automatically be `is dynamic`

This pragma is primarily intended for use with a future pragma export
mechanism, so that modules like P5tie can specify that they need access
to all variables in order to function correctly. This will also (though
it needs more work yet) provide a way to request `$_` also be dynamic.