Make Failure.mess robust when no message
Previously, if the exception lacked a `message` method, then it would
blow up with a method not found error when trying to display the
Failure. Since this can happen via DESTROY for an unhandled Failure,
this could happen at pretty much any point in the program. Make it
robust in the face of a missing `message` method in the exception.
Resolves #2623.