Hacky fix for R#2640
After spending most of the day yak shaving for this, decided to just
hack a solution for the release to get the pressure off.  I have a patch
that will fix the re-composing of classes with meta methods in them, but
that also unfortunately disables this hack.

As was observed, a class with a ^parameterize in them will smartmatch
correctly after having done a .^compose on them.  This commit will do that
at the end of compiling the setting for the classes in the setting that
have a ^parameterize method.  This will **not** fix any classes out there
that also have a ^parameterize method.