Allow CStruct pass-by-value in Nativecall
This is the initial implementation of the pass-by-value
functionality for nativecall, it works on Parameters by recognizing
an 'is copy' trait, and as a new 'is copy' trait on Routine for the
return value. The 'copy' key is then added to the parameter/return
hash. The patch includes also tests of the basic functionality.