Create working runners in build directory. Generate perl6-debug executable.
The debug runner fails when starting at the moment though. Not sure what
it is yet.
The previous commit did not manage to fix the executables during build,
because it relied on the binary perl6. That one sadly can't work in a build
environment because the folder structure of PERL6_HOME doesn't match. Thus
we rely on the old .bat runnersin the build dir again.