Use new local symbol debug name support
When we lower lexicals into locals, it makes them invisible to all that
uses the MoarVM debug server interface. This was already the case, but
became rather more serious of late, since we now lower many further
variables. With this change, we now record the mappings of registers
into debug names, which are on the MoarVM backend now passed along to
the VM for its use in debugging. Since we did lower some things prior to
this change, however, this not only keeps the debug experience from
getting worse with this release, but also makes it better than it was
before.