Hacky fix for R#2686
Compilation complained about IO::Glob not having a "flat" method.  This is
strange, as IO::Glob does Iterable, and the Iterable role provides a "flat"
method.

The "flat" method was called during the setting up of the export of "glob"
in Rakudo::Internals.  Since this code predates the GLR, and the intent seemed
to be to iterate over whatever is in @*PACKAGES, it felt appropriate to
replace the "flat" by a ".list".  This spectests clean and unbreaks IO::Glob.
As to the deeper reason of why this worked, I have no idea.