Fixed a bug when a wrapped method is called first
For wrapper-like dispatchers (with WrapDispatcher being currently the
only one) we won't set explicit dispatcher for the last candidate on the
queue effectively allowing a control call like callsame/nextsame to
obtain routine's default dispatcher. This fixes a situation when a
wrapped method cannot pass control over to the next method on
MethodDispatcher queue.

rakudo/rakudo#2178