Make a Failure throw when it is used as something Iterable
Fixes R#2650.  Makes .iterator throw immediately.  There was some discussion
on #perl6-dev on whether or not to have .iterator just return self, rather
than throwing.  But that would just delay the throwing until the FALLBACK of
Failure would catch the "pull-one", which felt as just wasting extra cycles.

In the past, code was changed for similar situations to throw the exception.
This change should allow us to actually just return the Failure.

This causes some breakage in S03-operators/minmax.t, as the test for failure
is now thrown earlier and the tests don't expect that.