Make SetHash iterators more safe
Currently, deleting a key from a nqp::iterator of an nqp::hash and then
continuing iterating over the hash, may cause segfaults.  Prevent that
situation by letting the SetHash.kv/pairs/values methods run on an iterator
on a pre-made list of keys, rather than directly on the nqp::hash iterator.

Deleting can happen if the value of a key from a SetHash is set to False:

    my %s is SetHash = <a b c d e>;
    dd %s;  # SetHash.new("d","b","e","c","a")
    $_ = False for %s.values;
    dd %s;  # SetHash.new()