Make BagHash iterators more safe
Currently, deleting a key from a nqp::iterator of an nqp::hash and then
continuing iterating over the hash, may cause segfaults.  Prevent that
situation by letting the BagHash.kv/pairs/values methods run on an iterator
on a pre-made list of keys, rather than directly on the nqp::hash iterator.

Deleting can happen if the value of a key from a BagHash is set to 0:

    my %b is BagHash = <a b b c c c>
    dd %b;  # ("b"=>2,"c"=>3,"a"=>1).BagHash
    $_ = 0 for %b.values;
    dd %b;  # ().BagHash