Make MixHash iterators more safe
Currently, deleting a key from a nqp::iterator of an nqp::hash and then
continuing iterating over the hash, may cause segfaults.  Prevent that
situation by letting the MixHash.kv/pairs/values methods run on an iterator
on a pre-made list of keys, rather than directly on the nqp::hash iterator.

Deleting can happen if the value of a key from a MixHash is set to 0:

    my %m is MixHash = <a b b c c c>
    dd %m;  # ("b"=>2,"c"=>3,"a"=>1).MixHash
    $_ = 0 for %m.values;
    dd %m;  # ().MixHash