Retain original provides metadata structure
Installing a distribution would change `provides => { a => "a.pm6" }`
to something like `provides => { a => { "a.pm6" => { ... } } }`,
with the purpose of saving some extra meta data about the file.

This removes the need for the additional meta data, and makes
installation no longer modify the provides structure. In the
future we may find the meta data useful again, but it should
be put into its own top level field.