Simplify candidate selection in MAIN usage
A recent change tried to be smarter on which MAIN candidates to show.  This
was however too smart for its own good, omitting candidates when it shouldn't.
Now limiting of candidates will only happen if the first given parameter
matches a **constant** value as the first parameter.  So if you have multiple
MAIN candidates that have "foo" as the first parameter, and you give "foo"
as the first command line argument, then only those candidates will be
selected for the usage message if no appropriate candidate was found for
execution.  In all other cases, all candidates will be shown if no candidate
was found for execution.

Spotted at Andrew Shitov's talk at the German Perl Workshop.