Make a single named param match @ sig in MAIN
Fixes R#2797.  This does **not** rewrite the signature binder.  Instead,
if the generated capture can not find any candidates, the capture changes
any named scalar values into 1 element arrays, and tries to find candidates
again.  If that is succesful, the modified signature is dispatched.

This is really a hack, but will in most cases do the right.