Initial support for macro functions.
A macro of form @macro(...)@ is not cosidered a macro function. It
defines a name of a method on NQP::Configure::Macros class which is the
macro name prefixed with _m_: _m_macro.

If a macro doesn't have a parameter the parenthesises are still
required to distiguish it from config variables.

Parameter may contain macros too. Nesting allowed:

@macro(Text @config_var@ @another_macro(@@config_var1@@)@ and something else.)@

How nesteds are expanded is yet to be decided.