Avoid hack involving early return in Test.pm6
Instead, implement it with an exception. This means that we could lift
the restriction on needing a `Callable` that can be returned from in
this place too; unfortunately, we have spectests that make sure it
fails, so simply lifting the restriction will cause failures in the
spectest for released language versions. For now, just preserve the
existing behavior, but in a way that won't break our more accurate
tracking of which things need return handlers.