Make buidl dir runners reliable wrt install directory
Change the order of the includes to prefer the ones in the build dir.
The install dir include is necessary, because the runners need the NQP
installation which usually resides in the same install location.
So the best thing to do is make sure the stuff in the build dir is included
first.
I tested this by corrupting blib/Perl6/BOOTSTRAP.moarvm and perl6.moarvm
in the install and build directory respectively to be sure which one is
included first.