Test reproducibility of precompilation
Compile NativeCall.pm6 twice and check that it produces the exact same bytecode
as signified by the precomp file's checksum. NativeCall is a good candidate for
this as it's a somewhat complex and good sized module with a couple of
dependencies. This should cover a nice range of reproducibility issues.