Work around reproducible-builds.t failing on first run
When run for the first time during make test, reproducible-build.t failed
because it would precompile NativeCall's dependencies on the first compilation
but of NativeCall but re-use them on the second compilation. Precompiling
those dependencies causes serialization of an additional Match object, making
the test fail.

Preferrably it wouldn't matter that we need to precompile the dependencies.
Until that's fixed, make the test pass by ensuring that NativeCalls's
dependencies are already precompiled before the first test run. We'll still
get reproducible builds for installation of modules which is where it's most
important.