Add support for bundling arguments to the default argument parser
This adds a new entry to %*SUB-MAIN-OPTS, bundling. If it is enabled,
any single dash argument will be interpreted as a bundle of single
character arguments: -ab will parse to :a, :b instead of :ab. If it is
not enabled, or if the argument starts with a double dash, it will
still parse as :ab as before.