Fix bogus role conflict when the same method is added under multiple names
It is possible to add the same method object under different names to a class
or role. Thus we must not rely on the method object's name but always use the
name given to add_method when processing lists of methods.

Many thanks to Xliff++ for the excellent report! Fixes GH #2873