Improve GNU C++ name mangling
This makes it so the symbol is not assumed to be a constructor when it
includes new unless it's the last part of the symbol and infers whether
or not the symbol is a constructor by checking if the last two parts of
it are identical.

This also mangles Str, Pointer, and CArray attributes with the rw or
cpp-const traits properly.