Fix compile_in_context with an immediate block
This situation could occur in the case of a begin-time EVAL that had a
code expression that contained an immediate QAST block. This led to the
failed $_ lookup reported in #2779, meaning that compilation of that
module no longer fails. Unfortunately, that isn't enough to make it
actually work, but it's a step closer.