Replace build runner script with a binary
This solves several problems related to argument passing with .bat files
on Windows. This is a part of the work to get Perl 6 working on Windows
with spaces in path. The problem already arises during the build when
precompiling modules. That calls a new perl6 process which on windows
happens to be a .bat file.