Finalized support for explicit and autodetected nqp locations.
Make it build when nqp is defined with --with-nqp and it's location is
different from --prefix. Though since such setup seems to be
non-functional Configure.pl would produce a warning. It is not a error
condition because perl6 builds and installs anyway.

If there are neither --prefix nor --with-nqp defined then nqp would be
looked for in PATH. If found then the prefix would be set based on the
nqp binary location. Backends will be chosen based on what nqp-<backend>
binaries are exist in that location.

rakudo/rakudo#29119