Install Rakudo NQP libs to Perl6 home
The Rakudo NQP files were previously installed into NQP_HOME/lib. That
creates problems when installing Rakudo and NQP into different prefixes.
In that case rakudo ends up putting the files into a separate folder from
NQP because of the different prefix. This in turn causes the concept of
NQP home and Perl6 home to fall apart, as the NQP libs in rakudos home are
neither in the NQP home folder, nor the Rakudo home folder.
We solve this by just installing Rakudos compiled NQP files into Perl6
home. The distinction is then not anymore "If it's written in NQP put it in
NQP home, if in Perl 6 into Perl6 home." but instead "If it's part of NQP
put it in NQP home, if it's part of Rakudo put it in Perl6 home." See
https://github.com/rakudo/rakudo/issues/2919#issuecomment-496406792
and following for the discussion surrounding this change.