Put the core Perl 6 CompUnit repo in <Perl6 home>/core
Before it resided directly in <Perl6 home>. That was kind of messy as other
folders also reside in <Perl6 home> and are thus mixed in between the
folders of the CompUnit repo. The stale files of previous installs will not
be automatically removed, but will do no harm either.
See
https://github.com/rakudo/rakudo/issues/2919#issuecomment-496417006
for the discussion surrounding this change.