Start initializing most attributes at creation
Rather than waiting for their first touch. As well as the expected cases
that we need to exclude (BUILD + default or required), we also need to
for now exclude non-P6opaque cases. This is actually a broad brush for
the storage of arrays and hashes; we rely on the auto-viv there too
(though will address that in future changes).

A handful of spectests are unhappy about this change, and will need a
closer look.