Remove harmful stale files
Files in NQP_HOME/lib/Perl6 left there of previous installs will be loaded
before files in PERL6_HOME/lib/Perl6 and thus have precedence. Because of
that these files break the install if left there.
We thus check for their existence and remove them during install if
necessary. The user is notified of this during Configure.