Use a static exec path with non-reloc builds
In the non-relocatable case it's not necessary to determine the executable
path at runtime. We can just as well compile it in. Doing so solves the
problem of OpenBSD where it's currently not technically possible to
determine the path of an executable from within.

Fixes #2904

Also some cleanup of the C runner.