Introducing/Using the CachedIterator role
This is CachedIterator role is a refinement of the PredictiveIterator role.
The difference is that a CachedIterator role is supposed to supply a "cache"
method that will be called to set the cache of the Seq as soon as the iterator
is being used.

By having R:Iterator.ReifiedArrayIterator and R:Iterator.ReifiedListIterator
consume this role, makes the following code work again:

    given Seq.new((1,2,3).iterator) {
        dd .elems; dd .all; dd .all
    }
    # 3
    # all(1, 2, 3)
    # all(1, 2, 3)

which was broken by:

  https://github.com/rakudo/rakudo/commit/9107215a99f106835a121d8446d97d11fc017758