Implement more fine-grained deconting of exported symbols
The problem with containerized exports has been narrowed down to EXPORT
sub using hash for declaring exports where all values are getting
wrapped into Scalars. To fix that method import on World got a named
parameter :need-decont which signals that imported symbols need to be
checked for their sigils. Unless a sigil is $ or & the symbol value is
nqp::decont'ed.

This commit works for both rakudo/rakudo#2979 and rakudo/rakudo#3012.