Fix "Missing serialize REPR function for REPR NativeRef (StrLexRef)"
ASN::META parses a definition with a grammar and creates classes and methods
from the parsed description. This leads to a StrLexRef getting used as a method
name which ends up in the serialization context. Decont the names of added
methods to ensure we have a plain old string.

Fixes GH #3045