Bump the margin of a performance test
According to the original issue one of the variants wasn't simply
slower but rather used to hang. To test for the hang, the margin can
be relatively large. Currently the difference is about 2x, which means
that the test flapped depending on the system's load. Bumping it to 5x
should be good enough in most of the cases.

Resolves GH-3065.